Change default for `i8` to sum to `i32`
Since `i16` really isn't that large, we changed the default sum for `i8` to `i32` in both `ScoreValue` and `ErrorValue`.

+++ packages/ec-core/src/performance/score_value.rs
@@ -219,7 +219,7 @@ unhindered_accumulate::default_to! {
     ScoreValue<u128> => KeepResults<SaturatingSum>,
     ScoreValue<usize> => KeepResults<SaturatingSum>,
 
-    ScoreValue<i8> => KeepResults<Widen<ScoreValue<i16>, SumStrategy>>,
+    ScoreValue<i8> => KeepResults<Widen<ScoreValue<i32>, SumStrategy>>,
     ScoreValue<i16> => KeepResults<Widen<ScoreValue<i32>, SumStrategy>>,
     ScoreValue<i32> => KeepResults<Widen<ScoreValue<i64>, SumStrategy>>,
     ScoreValue<i64> => KeepResults<Widen<ScoreValue<i128>, SumStrategy>>,
